Formatted types with prettier
