+++ src/main/webapp/resources/js/components/lists/BasicList.tsx
@@ -5,9 +5,25 @@
 
 import React from "react";
 import { List, Typography } from "antd";
-import { BasicListProps } from "./BasicList.types";
+import {GridProps} from "../../types/ant-design/ant-design";
+import {ListItemLayout} from "antd/lib/list";
 
 const { Text } = Typography;
+
+export type BasicListItem =
+    | { title: string; desc: number; props?: JSX.ElementAttributesProperty[] }
+    | {
+    title: string;
+    desc: JSX.Element;
+    props?: JSX.ElementAttributesProperty[];
+}
+    | { title: string; desc: string; props?: JSX.ElementAttributesProperty[] };
+
+export interface BasicListProps {
+    itemLayout?: ListItemLayout | undefined;
+    dataSource: BasicListItem[];
+    grid?: GridProps;
+}
 
 /**
  * Stateless UI component for displaying a basic list with a title and
