+++ src/main/webapp/resources/js/types/i18n.d.ts
