+++ src/main/webapp/resources/js/types/irida.window.d.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 interface IridaWindow extends Window {
-    project?: any;
-    translations?: Record<string, string>;
-    TL?: any;
-    PAGE?: any;
-    IRIDA?: any;
-    GALAXY?: any;
+  project?: any;
+  translations?: Record<string, string>;
+  TL?: any;
+  PAGE?: any;
+  IRIDA?: any;
+  GALAXY?: any;
 }
