+++ src/main/webapp/resources/js/types/irida/ExportUpoadState.ts
@@ -1,17 +1,17 @@
 export enum ExportUploadState {
-    NEW = "NEW",
-    UPLOADING = "UPLOADING",
-    UPLOADED = "UPLOADED",
-    UPLOADED_ERROR = "UPLOAD_ERROR",
-    CREATED = "created",
-    FAILED = "failed",
-    QUEUED = "queued",
-    PROCESSING = "processing",
-    PROCESSED_OK = "processed-ok",
-    PROCESSED_ERROR = "processed-error",
-    WAITING = "waiting",
-    SUBMITTED = "submitted",
-    DELETED = "deleted",
-    RETIRED = "retired",
-    UNKNOWN = "unknown",
+  NEW = "NEW",
+  UPLOADING = "UPLOADING",
+  UPLOADED = "UPLOADED",
+  UPLOADED_ERROR = "UPLOAD_ERROR",
+  CREATED = "created",
+  FAILED = "failed",
+  QUEUED = "queued",
+  PROCESSING = "processing",
+  PROCESSED_OK = "processed-ok",
+  PROCESSED_ERROR = "processed-error",
+  WAITING = "waiting",
+  SUBMITTED = "submitted",
+  DELETED = "deleted",
+  RETIRED = "retired",
+  UNKNOWN = "unknown",
 }
