+++ src/main/webapp/resources/js/types/webpack.d.ts
